Handle buffer overflow exceptions in StreamWriter
StreamWriter may raise a buffer over flow exception which is expected to be handled by callers.

If one occurs attempt to wait on the writers condition variable (which may well fail if the writer is not currently executing) or (if the wait does fail) simply sleep for .2 seconds before trying again.